+++ config/global_constants.py
@@ -1,10 +1,16 @@
 '''File to handle global constants and environment variable'''
 
+from logger.enums import LogLevel
+
 
 class GlobalConstants():
     
     # Miscellaneous
     CURRENT_TIME = ''
+
+    # Logging
+    LOG_LEVEL = LogLevel.INFO
+    LOG_FILE = ''
 
     # Mining Operation
     OPERATION_LENGTH = 0