+++ mining_operation/mining_operation.py
@@ -3,6 +3,8 @@
 import random
 
 from config.global_constants import GlobalConstants
+from logger.enums import LogLevel
+from logger.logger import Logger
 from mining_operation.enums import TruckStatus
 from mining_operation.report_generator import ReportGenerator
 
@@ -98,18 +100,23 @@ class MiningOperation():
 
         mining_trucks_count = len(self.mining_trucks)
         mining_stations_count = len(self.mining_stations)
-        print(f'OPERATION_LENGTH: {GlobalConstants.OPERATION_LENGTH}')
-        print(f'Mining Trucks count: {mining_trucks_count}')
-        print(f'Mining Stations count: {mining_stations_count}\n')
+
+        if GlobalConstants.LOG_LEVEL in [LogLevel.INFO, LogLevel.DEBUG]:
+            Logger.Log(f'\n==================================== Operation Started ====================================\n')
+            Logger.Log(f'Logging level: {GlobalConstants.LOG_LEVEL.name}')
+            Logger.Log(f'Operation length: {GlobalConstants.OPERATION_LENGTH}')
+            Logger.Log(f'Mining Trucks count: {mining_trucks_count}')
+            Logger.Log(f'Mining Stations count: {mining_stations_count}\n')
+
         assert not (mining_trucks_count <= 0 or \
             mining_stations_count <= 0 or \
             GlobalConstants.OPERATION_LENGTH <= 0), \
-            f'One of the key information is missing to start the operation!\nMining Truck Count = {mining_trucks_count}\nMining Stations Count = {mining_stations_count}\nOperation length = {GlobalConstants.OPERATION_LENGTH}'
+            Logger.Log(f'One of the key information is missing to start the operation!\nMining Truck Count = {mining_trucks_count}\nMining Stations Count = {mining_stations_count}\nOperation length = {GlobalConstants.OPERATION_LENGTH}')
         while self.elapsed_time <= GlobalConstants.OPERATION_LENGTH and not self.operation_ended:
 
             # Get the truck that will finish it's job next and when
-            truck_name = self.FindNextTruckToFinish()
-            time_forward = self.mining_trucks[truck_name][1]
+            truck_id = self.FindNextTruckToFinish()
+            time_forward = self.mining_trucks[truck_id][1]
 
             # Forward all the truck's time to that point
             self.ForwardTimeWith(time_forward)
@@ -117,12 +124,14 @@ class MiningOperation():
                 break
 
             # Update the truck's status that will finish it's job to the next job
-            if self.mining_trucks[truck_name][1] == 0:
-                self.UpdateTruckStatus(truck_name)
+            if self.mining_trucks[truck_id][1] == 0:
+                self.UpdateTruckStatus(truck_id)
 
         # Generate the reports for the operation
         ReportGenerator.GenerateForMiningTrucks(self.mining_trucks_history)
         ReportGenerator.GenerateForMiningStations(self.mining_stations)
+        if GlobalConstants.LOG_LEVEL in [LogLevel.INFO, LogLevel.DEBUG]:
+            Logger.Log(f'\n\n\n==================================== Operation Ended ====================================')
         return True
 
 
@@ -131,22 +140,21 @@ class MiningOperation():
 
         truck_data_to_finish_next = ()
         # print(f'mining_trucks: {self.mining_trucks}')
-        for truck_name in self.mining_trucks:
+        for truck_id in self.mining_trucks:
             # Extract the data of the truck
-            truck_data = self.mining_trucks[truck_name]
+            truck_data = self.mining_trucks[truck_id]
             time_left = truck_data[1]
             truck_status = truck_data[0]
-            # print(f'truck_name: {truck_name}')
-            # print(f'truck_status: {truck_status}')
-            # print(f'time_left: {time_left}')
-            # print(f'truck_data_to_finish_next: {truck_data_to_finish_next}')
+            if GlobalConstants.LOG_LEVEL in [LogLevel.DEBUG]:
+                Logger.Log(f'truck_id: {truck_id}, status: {truck_status}, time_left: {time_left}')
+                Logger.Log(f'truck_data_to_finish_next: {truck_data_to_finish_next}')
 
             if truck_data_to_finish_next == ():
                 # If it's the first truck save it as the base for comparison
-                truck_data_to_finish_next = (truck_name, time_left)
+                truck_data_to_finish_next = (truck_id, time_left)
             elif time_left < truck_data_to_finish_next[1]:
                 # Compare the time left with the saved one's, and if it's lower then switch it.
-                truck_data_to_finish_next = (truck_name, time_left)
+                truck_data_to_finish_next = (truck_id, time_left)
 
         if truck_data_to_finish_next != ():
             return truck_data_to_finish_next[0]
@@ -157,23 +165,25 @@ class MiningOperation():
     def ForwardTimeWith(self, time_forward:int):
         '''Method to forward the time for all the mining trucks'''
 
-        for truck_name in self.mining_trucks:
-            current_time = self.mining_trucks[truck_name][1]
+        for truck_id in self.mining_trucks:
+            current_time = self.mining_trucks[truck_id][1]
             new_time = current_time - time_forward
-            truck_status = self.mining_trucks[truck_name][0]
-            self.mining_trucks[truck_name] = (truck_status, new_time)
+            truck_status = self.mining_trucks[truck_id][0]
+            self.mining_trucks[truck_id] = (truck_status, new_time)
 
         # Increase the elapsed time to reflect the timeline correctly
         self.elapsed_time += time_forward
+        if GlobalConstants.LOG_LEVEL in [LogLevel.DEBUG]:
+            Logger.Log(f'\nTime forwarded: {time_forward}\n')
         if GlobalConstants.OPERATION_LENGTH <= self.elapsed_time:
             self.operation_ended = True
 
 
 
-    def UpdateTruckStatus(self, truck_name:str):
+    def UpdateTruckStatus(self, truck_id:str):
         '''Method to update the mining truck with the next job's data'''
 
-        status = self.mining_trucks[truck_name][0]
+        status = self.mining_trucks[truck_id][0]
 
         # If it was mining, switch to traveling and add the travel time
         if status is TruckStatus.MINING:
@@ -183,7 +193,7 @@ class MiningOperation():
         # If it was traveling, switch to queue up for mining station and add wait time
         elif status is TruckStatus.TRAVELING_TO_STATION:
             new_status = TruckStatus.WAITING_AT_STATION
-            new_value = self.QueueForStation(truck_name)
+            new_value = self.QueueForStation(truck_id)
 
         # If it was queued, switch to unloading and add the unloading time
         elif status is TruckStatus.WAITING_AT_STATION:
@@ -202,33 +212,42 @@ class MiningOperation():
         else:
             raise NotImplementedError(f'Status "{status}" handling not implemented!')
 
-        self.mining_trucks[truck_name] = (new_status, new_value)
-        self.mining_trucks_history = truck_name, (new_status, new_value)
+        if GlobalConstants.LOG_LEVEL in [LogLevel.DEBUG]:
+            Logger.Log(f'\nTruck #{truck_id} changing to: {new_status} for {new_value}\n')
+        self.mining_trucks[truck_id] = (new_status, new_value)
+        self.mining_trucks_history = truck_id, (new_status, new_value)
 
 
-    def QueueForStation(self, truck_name:str) -> int:
+    def QueueForStation(self, truck_id:str) -> int:
         '''Method to get the wait time for the truck at the station'''
 
         current_time = self.elapsed_time
         next_free_station = ()
         # Go through all the stations to find the one it can queue the truck for
-        for station_name in self.mining_stations:
-            if len(self.mining_stations[station_name]) == 0:
+        for station_id in self.mining_stations:
+            if len(self.mining_stations[station_id]) == 0:
                 latest_available_time = 0
             else:
-                latest_available_time = max(self.mining_stations[station_name])
+                latest_available_time = max(self.mining_stations[station_id])
 
+            if GlobalConstants.LOG_LEVEL in [LogLevel.DEBUG]:
+                Logger.Log(f'\nStation #{station_id} last unloading spot: {latest_available_time}')
+                Logger.Log(f'Current time: {current_time}')
             # If the last unloading finish is lower then the current time, then the station is free
             if latest_available_time < current_time:
-                next_free_station = (station_name, 0)
+                next_free_station = (station_id, 0)
+                if GlobalConstants.LOG_LEVEL in [LogLevel.DEBUG]:
+                    Logger.Log(f'Station #{station_id} is free')
                 break
             # If it wasn't free, but it's the first station, save the name and available time
             elif next_free_station == ():
-                next_free_station = (station_name, latest_available_time)
+                next_free_station = (station_id, latest_available_time)
             # If the available time is lower then the lowest previous station's,
             # then update which station will be free soonest
             elif latest_available_time < next_free_station[1]:
-                next_free_station = (station_name, latest_available_time)
+                next_free_station = (station_id, latest_available_time)
+            if GlobalConstants.LOG_LEVEL in [LogLevel.DEBUG]:
+                Logger.Log(f'Next free station {next_free_station}')
 
         # Calculate the wait time by distracting the current time from the latest available time
         if next_free_station[1] == 0:
@@ -237,14 +256,14 @@ class MiningOperation():
             wait_time = next_free_station[1] - current_time + 1
 
         # Add truck unloading time to the station
-        self.UpdateStationStatus(truck_name, next_free_station[0], wait_time)
+        self.UpdateStationStatus(truck_id, next_free_station[0], wait_time)
 
         if GlobalConstants.OPERATION_LENGTH < wait_time + current_time:
             self.operation_ended = True
         return wait_time
 
 
-    def UpdateStationStatus(self, truck_name:str, station_name:str, wait_time:int):
+    def UpdateStationStatus(self, truck_id:str, station_id:str, wait_time:int):
         '''Method to update the mining station with the incoming truck's data'''
 
         # Get the time stamp for the start and end time of the unloading of the truck at the station
@@ -255,7 +274,9 @@ class MiningOperation():
         for time_stamp in range(star_time, end_time):
             if GlobalConstants.OPERATION_LENGTH < time_stamp:
                 break
-            self.mining_stations[station_name][time_stamp] = truck_name
+            self.mining_stations[station_id][time_stamp] = truck_id
+            if GlobalConstants.LOG_LEVEL in [LogLevel.DEBUG]:
+                Logger.Log(f'Station #{station_id} reserving time slot {time_stamp} to truck #{truck_id}')
 
 
 def GetMiningLength(min_length:int=60, max_length:int=300) -> int: