+++ start.py
@@ -5,6 +5,7 @@ import sys
 import datetime
 
 from config.global_constants import GlobalConstants
+from logger.enums import LogLevel
 from mining_operation import MiningOperation
 
 
@@ -37,6 +38,12 @@ def ParseArguments():
                         default=4320,
                         type=int)
 
+    parser.add_argument('-l',
+                        help='The level of logging during code execution',
+                        required=False,
+                        default='INFO',
+                        type=str)
+
     args = parser.parse_args()
     return args
 
@@ -49,6 +56,8 @@ if __name__ == '__main__':
     # Setting global constants
     GlobalConstants.OPERATION_LENGTH = args.o
     GlobalConstants.CURRENT_TIME = datetime.datetime.now().strftime('%Y%m%d%H%M%S')
+    GlobalConstants.LOG_FILE = f'report_manual_{GlobalConstants.CURRENT_TIME}.log'
+    GlobalConstants.LOG_LEVEL = LogLevel.FromStr(args.l)
 
     operation = MiningOperation(
         mining_truck_count = args.n,